Make PRE wrap for the sake of less scrolling
Plain text emails are displaying with the PRE tag, which by default
does not wrap based on a given viewport width. To have less
side-scrolling to make it easier to read, even if it does show up
a lot more broken lines where the original line break cuts a line
in half.