Move menu bar to after body of email message
The aim is to read the message first and then afterwards, if there
is a reason to act, have it show up at the footer. Otherwise it would
be looking to act before really reading what the email is about.

Though it does move the jump link back to the main message link way
down the bottom as a result.